requirements.txt
-e git://github.com/bryanchow/python-jsconnect.git#egg=jsconnect

django-recaptcha==1.0.4
